Moving OAuth bearer tokens items
